refactor(proxy): extract optimizer stats handlers
Move optimizer statistics, range cardinality, and cost materialization handling into optimizer_stats.cc.
Register optimizer_stats.cc in the proxy build and keep THD context access behind handler helpers.
