+++ proxy/CMakeLists.txt
@@ -43,6 +43,7 @@ set(LINEAIRDB_SOURCES ha_lineairdb.cc ha_lineairdb.hh
                       index_scan.cc
                       table_scan.cc
                       multi_range_read.cc
+                      optimizer_stats.cc
                       lineairdb_field.cc lineairdb_field.hh
                       lineairdb_transaction.cc lineairdb_transaction.hh
                       lineairdb_proxy.cc lineairdb_proxy.hh
