+++ proxy/ha_lineairdb.cc
@@ -184,26 +184,6 @@ static void ensure_lineairdb_proxy(LineairDBThdCtx *&ctx) {
   }
 }
 
-/**
- * @brief List the indexes whose key-prefix NDV the server should measure.
- */
-static std::vector<std::pair<std::string, uint32_t>> index_ndv_descriptors(
-    TABLE *table) {
-  std::vector<std::pair<std::string, uint32_t>> descs;
-  if (table == nullptr || table->s == nullptr)
-    return descs;
-
-  descs.reserve(table->s->keys);
-  for (uint i = 0; i < table->s->keys; ++i) {
-    KEY *key = table->key_info + i;
-    const bool is_primary = (i == table->s->primary_key);
-    descs.emplace_back(is_primary ? std::string()
-                                  : std::string(key->name ? key->name : ""),
-                       key->user_defined_key_parts);
-  }
-  return descs;
-}
-
 static int lineairdb_commit(handlerton *hton, THD *thd, bool shouldCommit);
 static int lineairdb_abort(handlerton *hton, THD *thd, bool);
 
@@ -882,10 +862,6 @@ local_aggregate_fallback:;
   return false;
 }
 
-// Row count at or above which a grouped aggregate is served by the override
-// (a primary full scan) rather than a selective secondary index.
-static constexpr double kSmallAggregateInputRows = 1000.0;  // FIXME: make configurable
-
 /**
  * @brief Return true when the chosen leaf is a full PRIMARY scan.
  *
@@ -1204,333 +1180,6 @@ bool ha_lineairdb::agg_next_raw(std::string_view *out_value) {
 }
 
 /**
- * @brief Seed this table's row-count baseline from the proxy stats cache.
- */
-bool ha_lineairdb::seed_row_count_from_cache(LineairDBProxy *proxy) {
-  if (proxy == nullptr || share == nullptr || db_table_name.empty())
-    return false;
-
-  const auto &stats_cache = proxy->cached_table_stats();
-  auto it = stats_cache.find(db_table_name);
-  if (it == stats_cache.end() || it->second <= 0)
-    return false;
-
-  share->stats_base_records.store(static_cast<uint64_t>(it->second),
-                                  std::memory_order_relaxed);
-  for (auto &shard : share->rowcount_shards)
-    shard.delta.store(0, std::memory_order_relaxed);
-  return true;
-}
-
-/**
- * @brief Copy the proxy's latest per-index optimizer stats into shared table
- * metadata.
- */
-void ha_lineairdb::load_index_stats_from_cache(LineairDBProxy *proxy) {
-  if (proxy == nullptr || share == nullptr)
-    return;
-
-  const uint64_t records =
-      share->stats_base_records.load(std::memory_order_relaxed);
-  std::lock_guard<std::mutex> lock(share->index_ndv_mu_);
-  share->index_ndv_.clear();
-  share->index_hist_.clear();
-  for (const auto &entry : proxy->last_index_ndv()) {
-    if (entry.second.available)
-      share->index_ndv_[entry.first] = entry.second.values;
-  }
-  for (const auto &entry : proxy->last_index_hist()) {
-    const auto &hist = entry.second;
-    if (!hist.available || hist.bounds.empty() ||
-        hist.bounds.size() != hist.cum.size())
-      continue;
-
-    bool monotone = true;
-    for (size_t i = 1; i < hist.cum.size(); ++i) {
-      if (hist.cum[i] < hist.cum[i - 1] ||
-          hist.bounds[i] < hist.bounds[i - 1]) {
-        monotone = false;
-        break;
-      }
-    }
-    if (!monotone)
-      continue;
-
-    LineairDB_share::RangeHist range_hist;
-    range_hist.bounds = hist.bounds;
-    range_hist.cum = hist.cum;
-    share->index_hist_[entry.first] = std::move(range_hist);
-  }
-  share->index_ndv_records_.store(records, std::memory_order_relaxed);
-  share->index_ndv_loaded_.store(true, std::memory_order_relaxed);
-}
-
-/**
- * @brief Mark cached NDV stale when a SELECT sees more than 20% row-count
- * drift.
- */
-void ha_lineairdb::mark_stale_index_ndv_for_select() {
-  if (share == nullptr ||
-      !share->index_ndv_loaded_.load(std::memory_order_relaxed))
-    return;
-
-  THD *thd = ha_thd();
-  const bool is_select = thd != nullptr && thd->lex != nullptr &&
-                         thd->lex->sql_command == SQLCOM_SELECT;
-  if (!is_select)
-    return;
-
-  const uint64_t at_fetch =
-      share->index_ndv_records_.load(std::memory_order_relaxed);
-  const uint64_t now =
-      share->stats_base_records.load(std::memory_order_relaxed);
-  const uint64_t hi = std::max(at_fetch, now);
-  const uint64_t lo = std::min(at_fetch, now);
-  // Refetch when the gap is more than 20% of the larger row count.
-  if (hi == 0 || (hi - lo) * 5 <= hi)
-    return;
-
-  share->index_ndv_force_refresh_.store(true, std::memory_order_relaxed);
-  share->index_ndv_loaded_.store(false, std::memory_order_relaxed);
-}
-
-/**
- * @brief Fetch row-count and per-index NDV stats before optimizer planning.
- */
-void ha_lineairdb::seed_optimizer_stats() {
-  if (share == nullptr || db_table_name.empty())
-    return;
-
-  const bool need_rowcount =
-      share->stats_base_records.load(std::memory_order_relaxed) == 0;
-  mark_stale_index_ndv_for_select();
-  const bool need_ndv =
-      !share->index_ndv_loaded_.load(std::memory_order_relaxed);
-  const bool force_ndv =
-      share->index_ndv_force_refresh_.load(std::memory_order_relaxed);
-  if (!need_rowcount && !need_ndv && !force_ndv)
-    return;
-
-  THD *thd = ha_thd();
-  if (thd == nullptr)
-    return;
-
-  LineairDBThdCtx *&ctx = lineairdb_thd_ctx(thd, lineairdb_hton);
-  ensure_lineairdb_proxy(ctx);
-  if (ctx == nullptr || !ctx->proxy)
-    return;
-
-  // Existing path: use BEGIN/END piggyback stats when available.
-  const bool seeded = seed_row_count_from_cache(ctx->proxy.get());
-
-  // Cold optimizer paths can reach info() before tx_begin.
-  if (!seeded || need_ndv || force_ndv) {
-    bool fetched = false;
-    bool requested_ndv = false;
-    if (need_ndv || force_ndv) {
-      const auto descs = index_ndv_descriptors(table);
-      // Consume ANALYZE's force flag only when issuing the NDV RPC.
-      const bool force = share->index_ndv_force_refresh_.exchange(
-          false, std::memory_order_relaxed);
-      fetched = ctx->proxy->fetch_table_stats(db_table_name, descs, force);
-      requested_ndv = true;
-    } else {
-      fetched = ctx->proxy->fetch_table_stats();
-    }
-    if (fetched) {
-      seed_row_count_from_cache(ctx->proxy.get());
-      if (requested_ndv)
-        load_index_stats_from_cache(ctx->proxy.get());
-    }
-  }
-}
-
-/**
-  @brief
-  ::info() is used to return information to the optimizer. See my_base.h for
-  the complete description.
-
-  @details
-  Currently this table handler doesn't implement most of the fields really
-  needed. SHOW also makes use of this data.
-
-  You will probably want to have the following in your code:
-  @code
-  if (records < 2)
-    records = 2;
-  @endcode
-  The reason is that the server will optimize for cases of only a single
-  record. If, in a table scan, you don't know the number of records, it
-  will probably be better to set records to two so you can return as many
-  records as you need. Along with records, a few more variables you may wish
-  to set are:
-    records
-    deleted
-    data_file_length
-    index_file_length
-    delete_length
-    check_time
-  Take a look at the public variables in handler.h for more information.
-
-  Called in filesort.cc, ha_heap.cc, item_sum.cc, opt_sum.cc, sql_delete.cc,
-  sql_delete.cc, sql_derived.cc, sql_select.cc, sql_select.cc, sql_select.cc,
-  sql_select.cc, sql_select.cc, sql_show.cc, sql_show.cc, sql_show.cc,
-  sql_show.cc, sql_table.cc, sql_union.cc, and sql_update.cc.
-
-  @see
-  filesort.cc, ha_heap.cc, item_sum.cc, opt_sum.cc, sql_delete.cc,
-  sql_delete.cc, sql_derived.cc, sql_select.cc, sql_select.cc, sql_select.cc,
-  sql_select.cc, sql_select.cc, sql_show.cc, sql_show.cc, sql_show.cc,
-  sql_show.cc, sql_table.cc, sql_union.cc and sql_update.cc
-*/
-int ha_lineairdb::info(uint flag) {
-  DBUG_TRACE;
-
-  if (table == nullptr || table->s == nullptr) {
-    if (stats.records < 2)
-      stats.records = 2;
-    return 0;
-  }
-
-  if (flag & (HA_STATUS_VARIABLE | HA_STATUS_CONST)) {
-    seed_optimizer_stats();
-
-    int64_t delta_sum = 0;
-    for (const auto &shard : share->rowcount_shards) {
-      delta_sum += shard.delta.load(std::memory_order_relaxed);
-    }
-
-    const int64_t base = static_cast<int64_t>(
-        share->stats_base_records.load(std::memory_order_relaxed));
-    int64_t total = base + delta_sum;
-    if (total < 0)
-      total = 0;
-
-    stats.records = static_cast<ha_rows>(total);
-
-    // Check for uncommitted row-count delta in the active transaction
-    THD *thd = ha_thd();
-    if (thd != nullptr) {
-      LineairDBThdCtx *ctx =
-          *reinterpret_cast<LineairDBThdCtx **>(thd_ha_data(thd, lineairdb_hton));
-      LineairDBTransaction *active_tx = (ctx != nullptr) ? ctx->tx : nullptr;
-      if (active_tx != nullptr && !active_tx->is_not_started()) {
-        if (active_tx->is_aborted()) {
-          return abort_errno(active_tx);
-        }
-
-        const int64_t local_delta = active_tx->peek_rowcount_delta(share);
-        if (local_delta != 0) {
-          int64_t local_total =
-              static_cast<int64_t>(stats.records) + local_delta;
-          if (local_total < 0)
-            local_total = 0;
-          stats.records = static_cast<ha_rows>(local_total);
-        }
-      }
-    }
-
-    if (stats.records < 2)
-      stats.records = 2;
-
-    stats.mean_rec_length = table->s->reclength > 0 ? table->s->reclength : 100;
-    stats.data_file_length = stats.records * stats.mean_rec_length;
-    stats.index_file_length = stats.data_file_length / 2;
-  }
-  if ((flag & (HA_STATUS_CONST | HA_STATUS_VARIABLE)) && table != nullptr &&
-      table->s != nullptr) {
-    for (uint i = 0; i < table->s->keys; i++) {
-      KEY *key = table->key_info + i;
-      if (key == nullptr)
-        continue;
-      bool is_primary = (i == table->s->primary_key);
-      set_generic_rec_per_key(key, key->user_defined_key_parts, is_primary);
-    }
-  }
-
-  return 0;
-}
-
-int ha_lineairdb::analyze(THD *, HA_CHECK_OPT *) {
-  DBUG_TRACE;
-
-  if (share != nullptr) {
-    share->stats_base_records.store(0, std::memory_order_relaxed);
-    for (auto &shard : share->rowcount_shards)
-      shard.delta.store(0, std::memory_order_relaxed);
-    share->index_ndv_loaded_.store(false, std::memory_order_relaxed);
-    share->index_ndv_force_refresh_.store(true, std::memory_order_relaxed);
-  }
-
-  info(HA_STATUS_VARIABLE | HA_STATUS_CONST);
-  return HA_ADMIN_OK;
-}
-
-/**
- * @brief Estimate rec_per_key, the average rows matching each key prefix.
- *
- * @details Prefer server-measured NDV and compute rows / NDV(prefix). If NDV
- * is unavailable, keep the old uniform-distribution heuristic as a fallback.
- *
- * The fallback assumes N rows are split evenly across K key parts, so each
- * additional key part narrows the estimate by N^(1/K).
- *
- * Example with 1,000,000 rows and a 4-part unique key:
- * @verbatim
- * per_part = 1000000^(1/4) ~= 31.6
- * 1 part: 1000000 / 31.6   = 31,623 rows
- * 2 parts: 1000000 / 31.6^2 = 1,000 rows
- * 3 parts: 1000000 / 31.6^3 = 32 rows
- * 4 parts: full unique key  = 1 row
- * @endverbatim
- *
- * Full primary or unique keys are always costed as one row.
- *
- * See also: NDB's ndb_index_stat_set_rpk (ha_ndb_index_stat.cc:2529).
- */
-void ha_lineairdb::set_generic_rec_per_key(KEY *key, uint key_parts,
-                                           bool is_primary) {
-  bool is_unique = (key->flags & HA_NOSAME);
-
-  std::vector<uint64_t> ndv;
-  if (share != nullptr &&
-      share->index_ndv_loaded_.load(std::memory_order_relaxed)) {
-    const std::string index_name =
-        is_primary ? std::string() : std::string(key->name ? key->name : "");
-    std::lock_guard<std::mutex> lock(share->index_ndv_mu_);
-    auto it = share->index_ndv_.find(index_name);
-    if (it != share->index_ndv_.end() && it->second.size() >= key_parts)
-      ndv = it->second;
-  }
-  const bool has_ndv = ndv.size() >= key_parts;
-
-  // How much each additional key part narrows the result set (fallback path).
-  double per_part = std::max(
-      2.0, std::pow(static_cast<double>(stats.records), 1.0 / key_parts));
-
-  for (uint j = 0; j < key_parts; j++) {
-    ulong rpk; // records per key
-    if ((is_primary || is_unique) && j == key_parts - 1) {
-      // All parts specified on a UNIQUE/PK -> exactly 1 row
-      rpk = 1;
-    } else if (has_ndv && ndv[j] > 0) {
-      // Real stats: average rows per distinct key-prefix value.
-      const uint64_t records = static_cast<uint64_t>(stats.records);
-      const uint64_t distinct = ndv[j];
-      rpk = static_cast<ulong>(
-          std::max<uint64_t>(1, (records + distinct - 1) / distinct));
-    } else {
-      // per_part^(j+1) = total divisor for j+1 key parts
-      double selectivity = std::pow(per_part, static_cast<double>(j + 1));
-      rpk = static_cast<ulong>(
-          std::max(1.0, static_cast<double>(stats.records) / selectivity));
-    }
-    key->rec_per_key[j] = rpk;
-    key->set_records_per_key(j, static_cast<rec_per_key_t>(rpk));
-  }
-}
-
-/**
   @brief
   extra() is called whenever the server wishes to send a hint to
   the storage engine. The myisam engine implements the most hints.
@@ -1640,6 +1289,13 @@ int ha_lineairdb::start_stmt(THD *thd, thr_lock_type lock_type) {
   assert(lock_type > 0);
   userThread = thd;
   return external_lock(thd, lock_type);
+}
+
+LineairDBTransaction *ha_lineairdb::active_transaction(THD *thd) const {
+  if (thd == nullptr) return nullptr;
+  LineairDBThdCtx *ctx =
+      *reinterpret_cast<LineairDBThdCtx **>(thd_ha_data(thd, lineairdb_hton));
+  return (ctx != nullptr) ? ctx->tx : nullptr;
 }
 
 /**
@@ -1857,248 +1513,6 @@ int ha_lineairdb::rename_table(const char *, const char *, const dd::Table *,
                                dd::Table *) {
   DBUG_TRACE;
   return HA_ERR_WRONG_COMMAND;
-}
-
-/**
- * @brief Decode an integer key part from a MySQL range endpoint.
- *
- * @details Uses key_restore() into table->record[1] instead of reading raw key
- * bytes, so MySQL owns signedness and key-format decoding. Unsupported shapes
- * return false and the caller falls back to the old coarse estimate.
- */
-static bool helios_decode_keypart_int(TABLE *table, KEY *key, uint part,
-                                      const key_range *range,
-                                      longlong *out_value) {
-  if (table == nullptr || key == nullptr || range == nullptr ||
-      range->key == nullptr || out_value == nullptr) {
-    return false;
-  }
-  if (part >= key->user_defined_key_parts) return false;
-
-  uint required = 0;
-  for (uint i = 0; i <= part; ++i) {
-    required += key->key_part[i].store_length;
-  }
-  if (range->length < required) return false;
-
-  const KEY_PART_INFO &kp = key->key_part[part];
-  Field *field = kp.field;
-  if (field == nullptr || field->result_type() != INT_RESULT) return false;
-  if (field->is_nullable()) return false;
-  if (kp.key_part_flag & HA_REVERSE_SORT) return false;
-  if (table->record[0] == nullptr || table->record[1] == nullptr) return false;
-
-  uchar *scratch = table->record[1];
-  key_restore(scratch, range->key, key, range->length);
-  const ptrdiff_t delta =
-      static_cast<ptrdiff_t>(scratch - table->record[0]);
-  field->move_field_offset(delta);
-  *out_value = field->val_int();
-  field->move_field_offset(-delta);
-  return true;
-}
-
-/**
-  @brief
-  Given a starting key and an ending key, estimate the number of rows that
-  will exist between the two keys.
-
-  @details
-  end_key may be empty, in which case determine if start_key matches any rows.
-
-  Called from opt_range.cc by check_quick_keys().
-
-  @note
-  MySQL passes WHERE conditions as min_key / max_key:
-    - Equality (WHERE col=1):        min_key == max_key (same bytes)
-    - Range (WHERE col BETWEEN 1,5): min_key != max_key
-
-  We compare min_key and max_key byte-by-byte per key part to detect
-  how many leading columns are equality conditions, then use rec_per_key
-  for that depth. See NDB's records_in_range (ha_ndbcluster.cc:12861).
-
-  @see
-  check_quick_keys() in opt_range.cc
-*/
-ha_rows ha_lineairdb::records_in_range(uint inx, key_range *min_key,
-                                       key_range *max_key) {
-  DBUG_TRACE;
-
-  // Guard: table metadata not available
-  if (table == nullptr || table->s == nullptr) {
-    return 10;
-  }
-
-  // Get the index definition (inx = which index the optimizer is asking about)
-  KEY *key = table->key_info + inx;
-  if (key == nullptr) {
-    return 10;
-  }
-
-  // Total rows in this table (floor of 2 to avoid zero-cost estimates)
-  ha_rows total_records = stats.records;
-  if (total_records < 2)
-    total_records = 2;
-
-  // No bounds at all -> full table scan
-  if (min_key == nullptr && max_key == nullptr) {
-    return total_records;
-  }
-
-  // Count how many key parts (columns) the query specifies
-  uint key_parts_used = 0;
-  if (min_key != nullptr) {
-    key_parts_used = calculate_key_parts_from_length(key, min_key->length);
-  } else if (max_key != nullptr) {
-    key_parts_used = calculate_key_parts_from_length(key, max_key->length);
-  }
-
-  // All columns of a UNIQUE/PK specified -> exactly 1 row
-  if ((key->flags & HA_NOSAME) &&
-      key_parts_used == key->user_defined_key_parts) {
-    return 1;
-  }
-
-  // No columns matched -> full scan
-  if (key_parts_used == 0) {
-    return total_records;
-  }
-
-  // --- Equality prefix detection ---
-  // Compare min_key and max_key per column to find how many leading
-  // columns have equal values (= condition vs BETWEEN/range).
-  //
-  // Example: WHERE w=1 AND d=1 AND o_id BETWEEN 100 AND 200
-  //   min = [1][1][100], max = [1][1][200]
-  //     col 0: 1==1     -> eq
-  //     col 1: 1==1     -> eq
-  //     col 2: 100!=200 -> neq
-  //   -> eq_parts = 2
-  uint eq_parts = 0;
-  if (min_key != nullptr && max_key != nullptr) {
-    uint cmp_len = std::min(min_key->length, max_key->length);
-    uint consumed = 0; // byte offset into key buffer
-    for (uint p = 0; p < key->user_defined_key_parts && consumed < cmp_len; p++) {
-      uint part_len = key->key_part[p].store_length;
-      if (consumed + part_len > cmp_len)
-        break;
-      if (memcmp(min_key->key + consumed, max_key->key + consumed, part_len) == 0) {
-        eq_parts++;
-        consumed += part_len;
-      } else {
-        break; // This column differs -> range condition from here
-      }
-    }
-  } else if (min_key != nullptr) {
-    // One-sided: only min_key given. If exact match flag, treat as equality.
-    if (min_key->flag == HA_READ_KEY_EXACT ||
-        min_key->flag == HA_READ_KEY_OR_NEXT) {
-      eq_parts = key_parts_used;
-    }
-  }
-
-  // --- Estimate row count ---
-  ha_rows estimate;
-  if (eq_parts > 0) {
-    // Use rec_per_key at the equality depth
-    uint rpk_idx = eq_parts - 1; // 0-based array index
-    if (rpk_idx < key->user_defined_key_parts) {
-      estimate = static_cast<ha_rows>(key->rec_per_key[rpk_idx]);
-    } else {
-      estimate = 1;
-    }
-    // Range after an equality prefix: estimate how many distinct trailing
-    // values fall inside the range, then multiply by rows per trailing value.
-    // Unsupported key shapes fall back to the old /2 heuristic.
-    if (eq_parts < key_parts_used) {
-      longlong lo = 0;
-      longlong hi = 0;
-      if (eq_parts < key->user_defined_key_parts &&
-          key->rec_per_key[eq_parts] > 0 &&
-          helios_decode_keypart_int(table, key, eq_parts, min_key, &lo) &&
-          helios_decode_keypart_int(table, key, eq_parts, max_key, &hi) &&
-          hi >= lo) {
-        const double range_vals =
-            (hi > lo) ? static_cast<double>(hi - lo) : 1.0;
-        double est =
-            range_vals * static_cast<double>(key->rec_per_key[eq_parts]);
-
-        const double cap = static_cast<double>(key->rec_per_key[rpk_idx]);
-        if (cap >= 1.0 && est > cap) est = cap;
-        if (est < 1.0) est = 1.0;
-        estimate = static_cast<ha_rows>(est);
-      } else {
-        estimate = std::max(static_cast<ha_rows>(1), estimate / 2);
-      }
-    }
-  } else {
-    // No equality at all -> pure range scan.
-    // Heuristic: two-sided range ~5%, one-sided ~10% (NDB fallback).
-    if (min_key != nullptr && max_key != nullptr) {
-      estimate = std::max(static_cast<ha_rows>(2), total_records / 20);
-    } else {
-      estimate = std::max(static_cast<ha_rows>(2), total_records / 10);
-    }
-
-    // Leading-key range histogram: use the server-built cumulative counts as
-    // a floor over the heuristic. Missing or malformed stats keep the heuristic.
-    const bool is_primary = table->s != nullptr && inx == table->s->primary_key;
-    const std::string index_name =
-        is_primary ? std::string() : std::string(key->name ? key->name : "");
-    if (share != nullptr) {
-      std::lock_guard<std::mutex> lock(share->index_ndv_mu_);
-      auto hist_it = share->index_hist_.find(index_name);
-      if (hist_it != share->index_hist_.end()) {
-        const LineairDB_share::RangeHist &hist = hist_it->second;
-        if (!hist.bounds.empty() && hist.bounds.size() == hist.cum.size()) {
-          auto rank_le = [&](const std::string &encoded_key) -> double {
-            if (encoded_key >= hist.bounds.back())
-              return static_cast<double>(hist.cum.back());
-            auto it =
-                std::upper_bound(hist.bounds.begin(), hist.bounds.end(),
-                                 encoded_key);
-            if (it == hist.bounds.begin())
-              return 0.0;
-            const size_t pos =
-                static_cast<size_t>((it - hist.bounds.begin()) - 1);
-            return static_cast<double>(hist.cum[pos]);
-          };
-
-          constexpr key_part_map kLeadingPart = 1;
-          bool enc_ok = true;
-          double lo = 0.0;
-          double hi = static_cast<double>(hist.cum.back());
-          if (min_key != nullptr) {
-            std::string encoded = lineairdb_keyenc::convert_key_to_ldbformat(
-                table, inx, min_key->key, kLeadingPart);
-            if (encoded.empty())
-              enc_ok = false;
-            else
-              lo = rank_le(encoded);
-          }
-          if (enc_ok && max_key != nullptr) {
-            std::string encoded = lineairdb_keyenc::convert_key_to_ldbformat(
-                table, inx, max_key->key, kLeadingPart);
-            if (encoded.empty())
-              enc_ok = false;
-            else
-              hi = rank_le(encoded);
-          }
-          const double hist_est = enc_ok ? (hi - lo) : -1.0;
-          if (hist_est >= 1.0 &&
-              static_cast<ha_rows>(hist_est) > estimate) {
-            estimate = static_cast<ha_rows>(hist_est);
-          }
-        }
-      }
-    }
-  }
-
-  // Floor: always return at least 1 row to avoid zero-cost estimates
-  if (estimate < 1)
-    estimate = 1;
-
-  return estimate;
 }
 
 /**
@@ -2399,48 +1813,6 @@ bool ha_lineairdb::predict_prefetch_mode(THD *thd) {
       *reinterpret_cast<LineairDBThdCtx **>(thd_ha_data(thd, lineairdb_hton));
   if (ctx != nullptr && ctx->tx != nullptr) return ctx->tx->is_prefetch_mode();
   return srv_prefetch_execution && thd_can_use_prefetch(thd);
-}
-
-/**
- * @brief Return whether read_cost() should charge per-row materialization.
- *
- * @details Non-covering secondary ref/range reads usually fetch each base row
- * by PK after the index probe, so they should pay the materialization charge.
- * Primary-key reads already produce the base row, and grouped single-table
- * SELECTs served by bulk prefetch do not fetch rows one by one by PK.
- */
-bool ha_lineairdb::helios_charge_materialise(uint index, double rows) const {
-  const TABLE *t = table;
-  if (t == nullptr || t->in_use == nullptr) return true;
-
-  if (t->s != nullptr && t->s->primary_key != MAX_KEY &&
-      index == t->s->primary_key) {
-    return false;
-  }
-
-  THD *thd = t->in_use;
-  if (thd->lex == nullptr) return true;
-
-  // Only plain read-only SELECTs can use the bulk grouped path.
-  if (thd->lex->sql_command != SQLCOM_SELECT || thd->lex->is_explain())
-    return true;
-  if (t->reginfo.lock_type > TL_READ) return true;
-  if (!predict_prefetch_mode(thd)) return true;
-
-  // The grouped bulk path is recognizable before the final QEP exists.
-  Table_ref *tr = t->pos_in_table_list;
-  if (tr == nullptr || tr->query_block == nullptr) return true;
-  Query_block *qb = tr->query_block;
-  if (qb->leaf_table_count != 1 || !qb->is_grouped()) return true;
-  if (qb->is_distinct() || qb->olap != UNSPECIFIED_OLAP_TYPE ||
-      qb->has_windows()) {
-    return true;
-  }
-
-  // A large grouped aggregate is served by the override via a primary full scan.
-  // Above the threshold, charge so the optimizer prefers that primary scan;
-  // below it, keep the secondary cheap so a small aggregate still uses the index.
-  return rows >= kSmallAggregateInputRows;
 }
 
 struct st_mysql_storage_engine lineairdb_storage_engine = {
